LayoutTests/js/dom/domjit-accessor-node-type-effect-should-not-overlap-with-call-since-pure-expected.txt

Test function calls should not overlap with effects of DOMJIT's nodeType accessor since it is pure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test() is true
PASS successfullyParsed is true

TEST COMPLETE
